requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
Django==3.0.7
django-cors-headers==3.3.0
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.0.1
h11==0.14.0
idna==3.4
outcome==1.2.0
PyJWT==1.7.1
PySocks==1.7.1
pytz==2020.1
selenium==4.6.0
six==1.15.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.6
urllib3==1.26.12
wsproto==1.2.0
